Implement O(1) space k-group list reversal using modular pointer snapping and reconnecting logic

+++ Linked_List/Singly_Linked_List/Problem_Solving_Part/31_LC_143_Reorder_List.cpp
@@ -34,7 +34,7 @@ public:
         }
 
         int left = 0;
-        int right = nodes.size() - 1;
+        int right = static_cast<int>(nodes.size()) - 1;
 
         while (left < right)
         {
